pdf-saas/main-api-microservice/requirements.txt
python-multipart
python-dotenv
fastapi
uvicorn
aiofiles
httpx
pydantic